add basic data feild for room model
